Camera/requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
pillow>=10.0.0
requests>=2.31.0
duckduckgo-search>=6.2.0
ddgs>=9.0.0
